Add support for 'pending activity' in fidelity csv
After selling a position, you will have an extra "Pending activity"
item in the csv record that represents the value of the sold position.
This will be pending for a day or so, but can still be used to make
stock purchases, etc. For the purposes of this tool, lump this in with
the core position so that it is treated as uninvested capital that we
can use to bring our investments into alignment with the configured
target allocations.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>